+++ app/src/main/java/com/example/mainactivity/auth/firebase/FirebaseListener.java
@@ -1,4 +1,4 @@
-package com.example.mainactivity.login_signup.firebase;
+package com.example.mainactivity.auth.firebase;
 
 public interface FirebaseListener {
