+++ app/src/main/java/com/example/mainactivity/auth/login/LoginContract.java
@@ -1,4 +1,4 @@
-package com.example.mainactivity.login_signup.login;
+package com.example.mainactivity.auth.login;
 
 import android.content.Intent;
