+++ app/src/main/java/com/example/mainactivity/auth/login/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.mainactivity.login_signup.signup;
+package com.example.mainactivity.auth.login;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,26 +15,25 @@ import androidx.fragment.app.Fragment;
 
 import com.example.mainactivity.main.MainRouterContract;
 import com.example.mainactivity.R;
-import com.example.mainactivity.login_signup.firebase.FirebaseListener;
-import com.example.mainactivity.login_signup.firebase.FirebaseLoginSingleton;
+import com.example.mainactivity.auth.firebase.FirebaseListener;
+import com.example.mainactivity.auth.firebase.FirebaseLoginSingleton;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.textfield.TextInputEditText;
 
 import java.util.Objects;
 
-public class SignUpFragment extends Fragment implements FirebaseListener, SignUpContract.View {
+public class LoginFragment extends Fragment implements FirebaseListener, LoginContract.View {
 
-    Button btnSignUp, btnGoogleSignUp, btnFacebookSignUp;
+    Button btnLogin, btnGoogleLogin, btnFacebookLogin;
 
-    ConstraintLayout clFrSignUpScreen;
+    ConstraintLayout clFrLoginScreen;
 
-    EditText name;
     EditText email;
     TextInputEditText password;
 
     FirebaseLoginSingleton fbSingleton;
 
-    private final SignUpContract.Presenter presenter = new SignUpPresenter(this);
+    private final LoginContract.Presenter presenter = new LoginPresenter(this);
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
@@ -47,23 +46,22 @@ public class SignUpFragment extends Fragment implements FirebaseListener, SignUp
     public View onCreateView(@NonNull LayoutInflater inflater,
                              @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
+        View v = inflater.inflate(R.layout.fr_login_screen, container, false);
 
-        View v = inflater.inflate(R.layout.fr_sign_up_screen, container, false);
+        btnLogin = v.findViewById(R.id.btnLogin);
+        btnLogin.setOnClickListener(v1 -> presenter
+                .onLoginClicked(email.getText().toString(),
+                        Objects.requireNonNull(password.getText()).toString()));
+        btnGoogleLogin = v.findViewById(R.id.btnGoogleLogin);
+        btnGoogleLogin.setOnClickListener(v12 -> presenter.onLoginClickedGoogle());
+        btnFacebookLogin = v.findViewById(R.id.btnFacebookLogin);
+        btnFacebookLogin.setOnClickListener(v13 -> presenter
+                .onLoginClickedFacebook(LoginFragment.this));
 
-        name = v.findViewById(R.id.etName);
+        clFrLoginScreen = v.findViewById(R.id.clFrLoginScreen);
+
         email = v.findViewById(R.id.etEmail);
         password = v.findViewById(R.id.etPassword);
-
-        btnSignUp = v.findViewById(R.id.btnSignUp);
-        btnSignUp.setOnClickListener(v1 -> presenter.onSignUpClicked(name.getText().toString(),
-                email.getText().toString(), Objects.requireNonNull(password.getText()).toString()));
-        btnGoogleSignUp = v.findViewById(R.id.btnGoogleSignUp);
-        btnGoogleSignUp.setOnClickListener(v12 -> presenter.onSignUpClickedGoogle());
-        btnFacebookSignUp = v.findViewById(R.id.btnFacebookSignUp);
-        btnFacebookSignUp.setOnClickListener(v13 -> presenter
-                .onSignUpClickedFacebook(SignUpFragment.this));
-
-        clFrSignUpScreen = v.findViewById(R.id.clFrSignUpScreen);
 
         return v;
     }
@@ -76,31 +74,31 @@ public class SignUpFragment extends Fragment implements FirebaseListener, SignUp
 
     @Override
     public void onSuccess(String indicatorAuth) {
-        presenter.onSuccessSignUp(indicatorAuth);
+        presenter.onSuccessLogin(indicatorAuth);
     }
 
     @Override
     public void onFailure(String indicatorAuth) {
-        presenter.onFailureSignUp(indicatorAuth);
+        presenter.onFailureLogin(indicatorAuth);
     }
 
     @Override
-    public void signInWithEmail(String name, String email, String password) {
-        fbSingleton.executeSignUp(this, name, email, password);
+    public void signInWithEmail(String email, String password) {
+        fbSingleton.executeLogin(this, email, password);
     }
 
     @Override
     public void showSnackBar(String text) {
-        Snackbar.make(clFrSignUpScreen, text, Snackbar.LENGTH_SHORT).show();
+        Snackbar.make(clFrLoginScreen, text, Snackbar.LENGTH_SHORT).show();
     }
 
     @Override
-    public void startSignUpGoogle() {
+    public void startSignInGoogle() {
         fbSingleton.signIn();
     }
 
     @Override
-    public void startSignUpFacebook() {
+    public void startSignInFacebook() {
         fbSingleton.loginFacebook();
     }
 
@@ -115,12 +113,12 @@ public class SignUpFragment extends Fragment implements FirebaseListener, SignUp
     }
 
     @Override
-    public void successSignUp() {
+    public void successLogin() {
         ((MainRouterContract) requireActivity()).openHomeFragment();
     }
 
     @Override
-    public void failureSignUp(String text) {
-        Snackbar.make(clFrSignUpScreen, text, Snackbar.LENGTH_SHORT).show();
+    public void failureLogin(String text) {
+        Snackbar.make(clFrLoginScreen, text, Snackbar.LENGTH_SHORT).show();
     }
 }
