+++ app/src/main/java/com/example/mainactivity/main/MainRouterContract.java
@@ -2,11 +2,13 @@ package com.example.mainactivity.main;
 
 import android.os.Bundle;
 
+import com.example.mainactivity.retrofit.Post;
+
 public interface MainRouterContract {
 
     void openHomeFragment();
 
-    void openMoreAboutFragment(Bundle bundle);
+    void openMoreAboutFragment(Post post);
 
     void openSplashFragment();
