+++ app/src/main/java/com/example/mainactivity/more/AnswerMoreDetails.java
@@ -2,7 +2,9 @@ package com.example.mainactivity.more;
 
 import android.os.Bundle;
 
+import com.example.mainactivity.retrofit.Post;
+
 public interface AnswerMoreDetails {
 
-    void onSuccessMoreDetails(Bundle bundle);
+    void onSuccessMoreDetails(Post post);
 }
